fix module api
