+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './renderer/render';
+export * from './hooks/use-rendered';
 export * from './hooks/use-cluster';
 export * from './hooks/use-graphviz-context';
 export * from './hooks/use-digraph';
@@ -8,6 +9,7 @@ export * from './hooks/use-edge';
 export * from './hooks/use-node';
 export * from './hooks/use-rendered-id';
 export * from './hooks/use-root-cluster';
+export * from './components/Graphviz';
 export * from './components/HtmlLike';
 export * from './components/Graph';
 export * from './components/Digraph';
