--- src/web/index.ts
@@ -1,2 +1,0 @@
-export * from './hooks/rendered';
-export * from './components/Graphviz';
